Minor laoyt cleanup
